Graphical and logical additions:
* Link economy script to UI
* Add hoe and "X" sprites
* Replace old X in UI
* Link Hoe to gameobject sript

+++ Final2020/Assets/Scripts/Crop.cs
@@ -10,7 +10,7 @@ public class Crop : ScriptableObject
     public int growthState;
     public float growthTime;
     public bool done = false;
-    public float price;
+    public long price;
     public int count;
 
     private void Awake()
